user_service/src/requirements.txt
anyio==3.6.1
asyncpg==0.26.0
attrs==22.1.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
coverage==6.5.0
cryptography==38.0.1
databases==0.6.1
ecdsa==0.18.0
fastapi==0.85.1
greenlet==1.1.3.post0
h11==0.14.0
idna==3.4
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
psycopg2-binary==2.9.5
py==1.11.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.2
pyparsing==3.0.9
pytest==7.1.3
pytest-asyncio==0.20.1
pytest-cov==4.0.0
python-dotenv==0.21.0
python-jose==3.3.0
requests==2.28.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.42
SQLAlchemy-Utils==0.38.3
starlette==0.20.4
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.18.3
email-validator==1.3.0
